Implemented function findTasksByStatus

+++ src/main/java/com/example/ToDoList/DAO/StatusRepository.java
@@ -4,5 +4,5 @@ import com.example.ToDoList.DTO.StatusDto;
 import com.example.ToDoList.entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Integer> {
+public interface StatusRepository extends JpaRepository<Status, Long> {
 }
