+++ src/main/java/com/example/ToDoList/service/StatusService.java
@@ -1,20 +1,18 @@
 package com.example.ToDoList.service;
 
 import com.example.ToDoList.DTO.StatusDto;
-import com.example.ToDoList.entity.Status;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface StatusService {
 
     public List<StatusDto> getAllStatus();
 
-    public StatusDto findById(int id);
+    public StatusDto findById(Long id);
 
     public void save(StatusDto statusDto);
 
-    public void delete(int id);
+    public void delete(Long id);
 
-    void updateStatus(int id, StatusDto statusDto);
+    void updateStatus(Long id, StatusDto statusDto);
 }
